requirements.txt
# Основные библиотеки
python-telegram-bot==21.10  # Актуальная стабильная версия
requests==2.32.3  # Актуальная стабильная версия
python-dotenv==1.0.1  # Актуальная стабильная версия

# База данных
SQLAlchemy==2.0.38  # Актуальная стабильная версия

# Дата и время
python-dateutil==2.9.0.post0  # Актуальная стабильная версия

# Для работы с LLM
groq==0.18.0  # Актуальная стабильная версия

# Добавить если используете
aiosqlite==0.18.0

# Добавить для тестирования
pytest==7.3.1 